Crate linux_syscall

Linux syscall implementations

Example

The syscall is called like this in the zcore_loader:

let num = regs.rax as u32;
let args = [regs.rdi, regs.rsi, regs.rdx, regs.r10, regs.r8, regs.r9];
let mut syscall = Syscall {
    thread,
    thread_fn,
    syscall_entry: kernel_hal::context::syscall_entry as usize,
};
let ret = syscall.syscall(num, args).await;

Structs

Syscall

The struct of Syscall which stores the information about making a syscall