Struct linux_syscall::Syscall

pub struct Syscall<'a> {
    pub thread: &'a CurrentThread,
    pub syscall_entry: VirtAddr,
    pub regs: &'a mut GeneralRegs,
    pub thread_fn: ThreadFn,
}

The struct of Syscall which stores the information about making a syscall

Fields
thread: &'a CurrentThread

the thread making a syscall

syscall_entry: VirtAddr

the entry of current syscall

regs: &'a mut GeneralRegs

store the regs statues

thread_fn: ThreadFn

new thread function

Implementations

impl Syscall<'_>

pub fn sys_getcwd(&self, buf: UserOutPtr<u8>, len: usize) -> SysResult

return a null-terminated string containing an absolute pathname that is the current working directory of the calling process.

buf – pointer to buffer to receive path

len – size of buf

pub fn sys_chdir(&self, path: UserInPtr<u8>) -> SysResult

Change the current directory.

path – pointer to string with name of path

pub fn sys_mkdir(&self, path: UserInPtr<u8>, mode: usize) -> SysResult

Make a directory.

path – pointer to string with directory name

mode – file system permissions mode

pub fn sys_mkdirat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
mode: usize
) -> SysResult

create directory relative to directory file descriptor

pub fn sys_rmdir(&self, path: UserInPtr<u8>) -> SysResult

Remove a directory.

path – pointer to string with directory name

pub fn sys_getdents64(
&self,
fd: FileDesc,
buf: UserOutPtr<u8>,
buf_size: usize
) -> SysResult

get directory entries TODO: get ino from dirent

fd – file describe

pub fn sys_link(
&self,
oldpath: UserInPtr<u8>,
newpath: UserInPtr<u8>
) -> SysResult

creates a new link (also known as a hard link) to an existing file.

pub fn sys_linkat(
&self,
olddirfd: FileDesc,
oldpath: UserInPtr<u8>,
newdirfd: FileDesc,
newpath: UserInPtr<u8>,
flags: usize
) -> SysResult

create file link relative to directory file descriptors If the pathname given in oldpath is relative, then it is interpreted relative to the directory referred to by the file descriptor olddirfd

pub fn sys_unlink(&self, path: UserInPtr<u8>) -> SysResult

delete name/possibly file it refers to If that name was the last link to a file and no processes have the file open, the file is deleted. If the name was the last link to a file but any processes still have the file open, the file will remain in existence until the last file descriptor referring to it is closed.

pub fn sys_unlinkat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
flags: usize
) -> SysResult

remove directory entry relative to directory file descriptor The unlinkat() system call operates in exactly the same way as either unlink or rmdir.

pub fn sys_rename(
&self,
oldpath: UserInPtr<u8>,
newpath: UserInPtr<u8>
) -> SysResult

change name/location of file

pub fn sys_renameat(
&self,
olddirfd: FileDesc,
oldpath: UserInPtr<u8>,
newdirfd: FileDesc,
newpath: UserInPtr<u8>
) -> SysResult

rename file relative to directory file descriptors

pub fn sys_readlink(
&self,
path: UserInPtr<u8>,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

read value of symbolic link

pub fn sys_readlinkat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

read value of symbolic link relative to directory file descriptor readlink() places the contents of the symbolic link path in the buffer base, which has size len TODO: recursive link resolution and loop detection

impl Syscall<'_>

pub fn sys_open(
&self,
path: UserInPtr<u8>,
flags: usize,
mode: usize
) -> SysResult

Opens or creates a file, depending on the flags passed to the call. Returns an integer with the file descriptor.

pub fn sys_openat(
&self,
dir_fd: FileDesc,
path: UserInPtr<u8>,
flags: usize,
mode: usize
) -> SysResult

open file relative to directory file descriptor

pub fn sys_close(&self, fd: FileDesc) -> SysResult

Closes a file descriptor, so that it no longer refers to any file and may be reused.

pub fn sys_dup2(&self, fd1: FileDesc, fd2: FileDesc) -> SysResult

create a copy of the file descriptor oldfd.

pub fn sys_dup(&self, fd1: FileDesc) -> SysResult

create a copy of the file descriptor fd, and uses the lowest-numbered unused descriptor for the new descriptor.

pub fn sys_pipe(&self, fds: UserOutPtr<[i32; 2]>) -> SysResult

Creates a pipe, a unidirectional data channel that can be used for interprocess communication.

pub fn sys_pipe2(&self, fds: UserOutPtr<[i32; 2]>, flags: usize) -> SysResult

Creates a pipe, a unidirectional data channel that can be used for interprocess communication.

pub fn sys_flock(&mut self, fd: FileDesc, operation: usize) -> SysResult

apply or remove an advisory lock on an open file TODO: handle operation

impl Syscall<'_>

pub async fn sys_read(
&self,
fd: FileDesc,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

Reads from a specified file using a file descriptor. Before using this call, you must first obtain a file descriptor using the opensyscall. Returns bytes read successfully.

fd – file descriptor

base – pointer to the buffer to fill with read contents

len – number of bytes to read

pub fn sys_write(
&self,
fd: FileDesc,
base: UserInPtr<u8>,
len: usize
) -> SysResult

Writes to a specified file using a file descriptor. Before using this call, you must first obtain a file descriptor using the open syscall. Returns bytes written successfully.

fd – file descriptor

base – pointer to the buffer write

len – number of bytes to write

pub async fn sys_pread(
&self,
fd: FileDesc,
base: UserOutPtr<u8>,
len: usize,
offset: u64
) -> SysResult

read from or write to a file descriptor at a given offset reads up to count bytes from file descriptor fd at offset offset (from the start of the file) into the buffer starting at buf. The file offset is not changed.

pub fn sys_pwrite(
&self,
fd: FileDesc,
base: UserInPtr<u8>,
len: usize,
offset: u64
) -> SysResult

writes up to count bytes from the buffer starting at buf to the file descriptor fd at offset offset. The file offset is not changed.

pub async fn sys_readv(
&self,
fd: FileDesc,
iov_ptr: UserInPtr<IoVecOut>,
iov_count: usize
) -> SysResult

works just like read except that multiple buffers are filled. reads iov_count buffers from the file associated with the file descriptor fd into the buffers described by iov (“scatter input”)

pub fn sys_writev(
&self,
fd: FileDesc,
iov_ptr: UserInPtr<IoVecIn>,
iov_count: usize
) -> SysResult

works just like write except that multiple buffers are written out. writes iov_count buffers of data described by iov to the file associated with the file descriptor fd (“gather output”).

pub fn sys_lseek(&self, fd: FileDesc, offset: i64, whence: u8) -> SysResult

repositions the offset of the open file associated with the file descriptor fd to the argument offset according to the directive whence

pub fn sys_truncate(&self, path: UserInPtr<u8>, len: usize) -> SysResult

cause the regular file named by path to be truncated to a size of precisely length bytes.

pub fn sys_ftruncate(&self, fd: FileDesc, len: usize) -> SysResult

cause the regular file referenced by fd to be truncated to a size of precisely length bytes.

pub async fn sys_sendfile(
&self,
out_fd: FileDesc,
in_fd: FileDesc,
offset_ptr: UserInOutPtr<u64>,
count: usize
) -> SysResult

copies data between one file descriptor and another.

pub async fn sys_copy_file_range(
&self,
in_fd: FileDesc,
in_offset: UserInOutPtr<u64>,
out_fd: FileDesc,
out_offset: UserInOutPtr<u64>,
count: usize,
flags: usize
) -> SysResult

copies data between one file descriptor and anothe, read from specified offset and write new offset back

pub fn sys_sync(&self) -> SysResult

causes all buffered modifications to file metadata and data to be written to the underlying file systems.

pub fn sys_fsync(&self, fd: FileDesc) -> SysResult

transfers (“flushes”) all modified in-core data of (i.e., modified buffer cache pages for) the file referred to by the file descriptor fd to the disk device

pub fn sys_fdatasync(&self, fd: FileDesc) -> SysResult

is similar to fsync(), but does not flush modified metadata unless that metadata is needed

pub fn sys_ioctl(
&self,
fd: FileDesc,
request: usize,
arg1: usize,
arg2: usize,
arg3: usize
) -> SysResult

Set parameters of device files.

pub fn sys_fcntl(&self, fd: FileDesc, cmd: usize, arg: usize) -> SysResult

Manipulate a file descriptor.

cmd – cmd flag

arg – additional parameters based on cmd

pub fn sys_access(&self, path: UserInPtr<u8>, mode: usize) -> SysResult

Checks whether the calling process can access the file pathname

pub fn sys_faccessat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
mode: usize,
flags: usize
) -> SysResult

Check user’s permissions of a file relative to a directory file descriptor TODO: check permissions based on uid/gid

pub fn sys_utimensat(
&mut self,
dirfd: FileDesc,
pathname: UserInPtr<u8>,
times: UserInOutPtr<[TimeSpec; 2]>,
flags: usize
) -> SysResult

change file timestamps with nanosecond precision

impl Syscall<'_>

pub async fn sys_poll(
&mut self,
ufds: UserInOutPtr<PollFd>,
nfds: usize,
timeout_msecs: usize
) -> SysResult

Wait for some event on a file descriptor

pub async fn sys_ppoll(
&mut self,
ufds: UserInOutPtr<PollFd>,
nfds: usize,
timeout: UserInPtr<TimeSpec>
) -> SysResult

Wait for some event on a file descriptor

ppoll() allows an application to safely wait until either a file descriptor becomes ready or until a signal is caught

pub async fn sys_pselect6(
&mut self,
nfds: usize,
read: UserInOutPtr<u32>,
write: UserInOutPtr<u32>,
err: UserInOutPtr<u32>,
timeout: UserInPtr<TimeVal>,
_sigset: usize
) -> SysResult

similar to select, but have sigmask argument

pub async fn sys_select(
&mut self,
nfds: usize,
read: UserInOutPtr<u32>,
write: UserInOutPtr<u32>,
err: UserInOutPtr<u32>,
timeout: UserInPtr<TimeVal>
) -> SysResult

allow a program to monitor multiple file descriptors, waiting until one or more of the file descriptors become “ready” for some class of I/O operation.

A file descriptor is considered ready if it is possible to perform the corresponding I/O operation (e.g., read) without blocking.

impl Syscall<'_>

pub fn sys_lstat(
&self,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>
) -> SysResult

Works exactly like the stat syscall, but if the file in question is a symbolic link, information on the link is returned rather than its target.

path – full path to file

stat_ptr – pointer to stat buffer

pub fn sys_fstat(&self, fd: FileDesc, stat_ptr: UserOutPtr<Stat>) -> SysResult

Works exactly like the stat syscall except a file descriptor (fd) is provided instead of a path.

fd – file descriptor

stat_ptr – pointer to stat buffer

pub fn sys_fstatat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>,
flags: usize
) -> SysResult

get file status relative to a directory file descriptor

pub fn sys_stat(
&self,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>
) -> SysResult

Returns information about a file in a structure named stat.

path – pointer to the name of the file

stat_ptr – pointer to the structure to receive file information

impl Syscall<'_>

pub fn sys_semget(&self, key: usize, nsems: usize, flags: usize) -> SysResult

returns the semaphore set identifier associated with the argument key

pub async fn sys_semop(
&self,
id: usize,
ops: UserInPtr<SemBuf>,
num_ops: usize
) -> SysResult

semaphore operations

performs operations on selected semaphores in the set indicated by semid

pub fn sys_semctl(
&self,
id: usize,
num: usize,
cmd: usize,
arg: usize
) -> SysResult

semaphore control operations

performs the control operation specified by cmd on the semaphore set identified by semid, or on the semnum-th semaphore of that set.

pub fn sys_shmget(&self, key: usize, size: usize, shmflg: usize) -> SysResult

allocates a shared memory segment

returns the identifier of the shared memory segment associated with the value of the argument key

pub fn sys_shmat(&self, id: usize, addr: VirtAddr, shmflg: usize) -> SysResult

attaches the shared memory segment identified by shmid to the address space of the calling process.

pub fn sys_shmdt(&self, id: usize, addr: VirtAddr, shmflg: usize) -> SysResult

detaches the shared memory segment located at the address specified by shmaddr from the address space of the calling process.

pub fn sys_shmctl(&self, id: usize, cmd: usize, buffer: usize) -> SysResult

shared memory control

performs the control operation specified by cmd on the shared memory segment whose identifier is given in id

impl Syscall<'_>

pub fn sys_arch_prctl(&mut self, code: i32, addr: usize) -> SysResult

set architecture-specific thread state for x86_64 currently

pub fn sys_uname(&self, buf: UserOutPtr<u8>) -> SysResult

get name and information about current kernel

pub fn sys_sysinfo(&mut self, sys_info: UserOutPtr<SysInfo>) -> SysResult

provides a simple way of getting overall system statistics

pub async fn sys_futex(
&self,
uaddr: usize,
op: u32,
val: i32,
timeout: UserInPtr<TimeSpec>
) -> SysResult

provides a method for waiting until a certain condition becomes true.

uaddr - points to the futex word.

op - the operation to perform on the futex

val - a value whose meaning and purpose depends on op

timeout - not support now TODO: support timeout

pub fn sys_prlimit64(
&mut self,
pid: usize,
resource: usize,
new_limit: UserInPtr<RLimit>,
old_limit: UserOutPtr<RLimit>
) -> SysResult

Combines and extends the functionality of setrlimit() and getrlimit()

pub fn sys_getrandom(
&mut self,
buf: UserOutPtr<u8>,
len: usize,
flag: u32
) -> SysResult

fills the buffer pointed to by buf with up to buflen random bytes.

buf - buffer that needed to fill

buflen - length of buffer

flag - a bit mask that can contain zero or more of the following values ORed together:

GRND_RANDOM

GRND_NONBLOCK

returns the number of bytes that were copied to the buffer buf.

impl Syscall<'_>

pub fn sys_rt_sigaction(
&self,
signum: usize,
act: UserInPtr<SignalAction>,
oldact: UserOutPtr<SignalAction>,
sigsetsize: usize
) -> SysResult

Used to change the action taken by a process on receipt of a specific signal.

pub fn sys_rt_sigprocmask(
&mut self,
how: i32,
set: UserInPtr<Sigset>,
oldset: UserOutPtr<Sigset>,
sigsetsize: usize
) -> SysResult

Used to fetch and/or change the signal mask of the calling thread

pub fn sys_sigaltstack(
&self,
ss: UserInPtr<SignalStack>,
old_ss: UserOutPtr<SignalStack>
) -> SysResult

Allows a process to define a new alternate signal stack and/or retrieve the state of an existing alternate signal stack

impl Syscall<'_>

pub fn sys_fork(&self) -> SysResult

Fork the current process. Return the child’s PID.

pub async fn sys_vfork(&self) -> SysResult

creates a child process of the calling process, similar to fork but wait for execve

pub fn sys_clone(
&self,
flags: usize,
newsp: usize,
parent_tid: UserOutPtr<i32>,
child_tid: UserOutPtr<i32>,
newtls: usize
) -> SysResult

Create a new thread in the current process. The new thread’s stack pointer will be set to newsp, and thread pointer will be set to newtls. The child tid will be stored at both parent_tid and child_tid. This is partially implemented for musl only.

pub async fn sys_wait4(
&self,
pid: i32,
wstatus: UserOutPtr<i32>,
options: u32
) -> SysResult

Wait for a child process exited.

Return the PID. Store exit code to wstatus if it’s not null.

pub fn sys_execve(
&mut self,
path: UserInPtr<u8>,
argv: UserInPtr<UserInPtr<u8>>,
envp: UserInPtr<UserInPtr<u8>>
) -> SysResult

Replaces the current ** process ** with a new process image

argv is an array of argument strings passed to the new program. envp is an array of strings, conventionally of the form key=value, which are passed as environment to the new program.

NOTICE: argv & envp can not be NULL (different from Linux)

NOTICE: for multi-thread programs A call to any exec function from a process with more than one thread shall result in all threads being terminated and the new executable image being loaded and executed.

pub fn sys_gettid(&self) -> SysResult

Get the current thread ID.

pub fn sys_getpid(&self) -> SysResult

Get the current process ID.

pub fn sys_getppid(&self) -> SysResult

Get the parent process ID.

pub fn sys_exit(&mut self, exit_code: i32) -> SysResult

Exit the current thread

pub fn sys_exit_group(&mut self, exit_code: i32) -> SysResult

Exit the current thread group (i.e. process)

pub async fn sys_nanosleep(&self, req: UserInPtr<TimeSpec>) -> SysResult

Allows the calling thread to sleep for an interval specified with nanosecond precision

pub fn sys_set_tid_address(&self, tidptr: UserOutPtr<i32>) -> SysResult

set pointer to thread ID returns the caller’s thread ID

impl Syscall<'_>

pub fn sys_clock_gettime(
&self,
clock: usize,
buf: UserOutPtr<TimeSpec>
) -> SysResult

finds the resolution (precision) of the specified clock clockid, and, if buffer is non-NULL, stores it in the struct timespec pointed to by buffer

pub fn sys_gettimeofday(
&mut self,
tv: UserOutPtr<TimeVal>,
tz: UserInPtr<u8>
) -> SysResult

get the time with second and microseconds

pub fn sys_time(&mut self, time: UserOutPtr<u64>) -> SysResult

get time in seconds

pub fn sys_getrusage(
&mut self,
who: usize,
rusage: UserOutPtr<RUsage>
) -> SysResult

get resource usage currently only support ru_utime and ru_stime:

ru_utime: user CPU time used

ru_stime: system CPU time used

pub fn sys_times(&mut self, buf: UserOutPtr<Tms>) -> SysResult

stores the current process times in the struct tms that buf points to

impl Syscall<'_>

pub async fn sys_mmap(
&self,
addr: usize,
len: usize,
prot: usize,
flags: usize,
fd: FileDesc,
offset: u64
) -> SysResult

creates a new mapping in the virtual address space of the calling process.

addr - The starting address for the new mapping

len - specifies the length of the mapping

prot - describes the desired memory protection of the mapping

flags - determines whether updates to the mapping are visible to other processes mapping the same region, and whether updates are carried through to the underlying file.

fd - mapping file descriptor

offset - offset in the file

pub fn sys_mprotect(&self, addr: usize, len: usize, prot: usize) -> SysResult

changes the access protections for the calling process’s memory pages containing any part of the address range in the interval [addr, addr+len-1] TODO: unimplemented

pub fn sys_munmap(&self, addr: usize, len: usize) -> SysResult

Deletes the mappings for the specified address range, and causes further references to addresses within the range to generate invalid memory references.

impl Syscall<'_>

pub async fn syscall(&mut self, num: u32, args: [usize; 6]) -> isize

syscall entry function