Struct linux_syscall::Syscall

pub struct Syscall<'a> {
    pub thread: &'a Arc<Thread>,
    pub syscall_entry: VirtAddr,
    pub regs: &'a mut GeneralRegs,
    pub spawn_fn: fn(thread: Arc<Thread>),
    pub exit: bool,
}

Fields
thread: &'a Arc<Thread>syscall_entry: VirtAddrregs: &'a mut GeneralRegsspawn_fn: fn(thread: Arc<Thread>)exit: bool

Set true to exit current task.

Implementations

impl<'_> Syscall<'_>

pub fn sys_getcwd(&self, buf: UserOutPtr<u8>, len: usize) -> SysResult

pub fn sys_chdir(&self, path: UserInPtr<u8>) -> SysResult

pub fn sys_mkdir(&self, path: UserInPtr<u8>, mode: usize) -> SysResult

pub fn sys_mkdirat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
mode: usize
) -> SysResult

pub fn sys_rmdir(&self, path: UserInPtr<u8>) -> SysResult

pub fn sys_getdents64(
&self,
fd: FileDesc,
buf: UserOutPtr<u8>,
buf_size: usize
) -> SysResult

pub fn sys_link(
&self,
oldpath: UserInPtr<u8>,
newpath: UserInPtr<u8>
) -> SysResult

pub fn sys_linkat(
&self,
olddirfd: FileDesc,
oldpath: UserInPtr<u8>,
newdirfd: FileDesc,
newpath: UserInPtr<u8>,
flags: usize
) -> SysResult

pub fn sys_unlink(&self, path: UserInPtr<u8>) -> SysResult

pub fn sys_unlinkat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
flags: usize
) -> SysResult

pub fn sys_rename(
&self,
oldpath: UserInPtr<u8>,
newpath: UserInPtr<u8>
) -> SysResult

pub fn sys_renameat(
&self,
olddirfd: FileDesc,
oldpath: UserInPtr<u8>,
newdirfd: FileDesc,
newpath: UserInPtr<u8>
) -> SysResult

pub fn sys_readlink(
&self,
path: UserInPtr<u8>,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

pub fn sys_readlinkat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

impl<'_> Syscall<'_>

pub fn sys_open(
&self,
path: UserInPtr<u8>,
flags: usize,
mode: usize
) -> SysResult

pub fn sys_openat(
&self,
dir_fd: FileDesc,
path: UserInPtr<u8>,
flags: usize,
mode: usize
) -> SysResult

pub fn sys_close(&self, fd: FileDesc) -> SysResult

pub fn sys_dup2(&self, fd1: FileDesc, fd2: FileDesc) -> SysResult

impl<'_> Syscall<'_>

pub fn sys_read(
&self,
fd: FileDesc,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

pub fn sys_write(
&self,
fd: FileDesc,
base: UserInPtr<u8>,
len: usize
) -> SysResult

pub fn sys_pread(
&self,
fd: FileDesc,
base: UserOutPtr<u8>,
len: usize,
offset: u64
) -> SysResult

pub fn sys_pwrite(
&self,
fd: FileDesc,
base: UserInPtr<u8>,
len: usize,
offset: u64
) -> SysResult

pub fn sys_readv(
&self,
fd: FileDesc,
iov_ptr: UserInPtr<IoVecOut>,
iov_count: usize
) -> SysResult

pub fn sys_writev(
&self,
fd: FileDesc,
iov_ptr: UserInPtr<IoVecIn>,
iov_count: usize
) -> SysResult

pub fn sys_lseek(&self, fd: FileDesc, offset: i64, whence: u8) -> SysResult

pub fn sys_truncate(&self, path: UserInPtr<u8>, len: usize) -> SysResult

pub fn sys_ftruncate(&self, fd: FileDesc, len: usize) -> SysResult

pub fn sys_sendfile(
&self,
out_fd: FileDesc,
in_fd: FileDesc,
offset_ptr: UserInOutPtr<u64>,
count: usize
) -> SysResult

pub fn sys_copy_file_range(
&self,
in_fd: FileDesc,
in_offset: UserInOutPtr<u64>,
out_fd: FileDesc,
out_offset: UserInOutPtr<u64>,
count: usize,
flags: usize
) -> SysResult

pub fn sys_sync(&self) -> SysResult

pub fn sys_fsync(&self, fd: FileDesc) -> SysResult

pub fn sys_fdatasync(&self, fd: FileDesc) -> SysResult

pub fn sys_ioctl(
&self,
fd: FileDesc,
request: usize,
arg1: usize,
arg2: usize,
arg3: usize
) -> SysResult

pub fn sys_fcntl(&self, fd: FileDesc, cmd: usize, arg: usize) -> SysResult

pub fn sys_access(&self, path: UserInPtr<u8>, mode: usize) -> SysResult

pub fn sys_faccessat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
mode: usize,
flags: usize
) -> SysResult

impl<'_> Syscall<'_>

pub fn sys_lstat(
&self,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>
) -> SysResult

pub fn sys_fstat(&self, fd: FileDesc, stat_ptr: UserOutPtr<Stat>) -> SysResult

pub fn sys_fstatat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>,
flags: usize
) -> SysResult

pub fn sys_stat(
&self,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>
) -> SysResult

impl<'_> Syscall<'_>

pub fn sys_arch_prctl(&mut self, code: i32, addr: usize) -> SysResult

pub fn sys_uname(&self, buf: UserOutPtr<u8>) -> SysResult

pub async fn sys_futex<'_>(
&'_ self,
uaddr: usize,
op: u32,
val: i32,
timeout: UserInPtr<TimeSpec>
) -> SysResult

impl<'_> Syscall<'_>

pub fn sys_fork(&self) -> SysResult

Fork the current process. Return the child's PID.

pub async fn sys_vfork<'_>(&'_ self) -> SysResult

pub fn sys_clone(
&self,
flags: usize,
newsp: usize,
parent_tid: UserOutPtr<i32>,
child_tid: UserOutPtr<i32>,
newtls: usize
) -> SysResult

Create a new thread in the current process. The new thread's stack pointer will be set to newsp, and thread pointer will be set to newtls. The child tid will be stored at both parent_tid and child_tid. This is partially implemented for musl only.

pub async fn sys_wait4<'_>(
&'_ self,
pid: i32,
__arg2: UserOutPtr<i32>,
options: u32
) -> SysResult

Wait for a child process exited.

Return the PID. Store exit code to wstatus if it's not null.

pub fn sys_execve(
&mut self,
path: UserInPtr<u8>,
argv: UserInPtr<UserInPtr<u8>>,
envp: UserInPtr<UserInPtr<u8>>
) -> SysResult

Replaces the current ** process ** with a new process image

argv is an array of argument strings passed to the new program. envp is an array of strings, conventionally of the form key=value, which are passed as environment to the new program.

NOTICE: argv & envp can not be NULL (different from Linux)

NOTICE: for multi-thread programs A call to any exec function from a process with more than one thread shall result in all threads being terminated and the new executable image being loaded and executed.

pub fn sys_gettid(&self) -> SysResult

Get the current thread ID.

pub fn sys_getpid(&self) -> SysResult

Get the current process ID.

pub fn sys_getppid(&self) -> SysResult

Get the parent process ID.

pub fn sys_exit(&mut self, exit_code: i32) -> SysResult

Exit the current thread

pub fn sys_exit_group(&mut self, exit_code: i32) -> SysResult

Exit the current thread group (i.e. process)

pub fn sys_set_tid_address(&self, tidptr: UserOutPtr<i32>) -> SysResult

impl<'_> Syscall<'_>

pub fn sys_clock_gettime(
&self,
clock: usize,
buf: UserOutPtr<TimeSpec>
) -> SysResult

impl<'_> Syscall<'_>

pub fn sys_mmap(
&self,
addr: usize,
len: usize,
prot: usize,
flags: usize,
fd: FileDesc,
offset: u64
) -> SysResult

pub fn sys_mprotect(&self, addr: usize, len: usize, prot: usize) -> SysResult

pub fn sys_munmap(&self, addr: usize, len: usize) -> SysResult

impl<'_> Syscall<'_>

pub async fn syscall<'_>(&'_ mut self, num: u32, args: [usize; 6]) -> isize