Struct linux_syscall::Syscall

pub struct Syscall<'a> {
    pub thread: &'a CurrentThread,
    pub thread_fn: ThreadFn,
    pub syscall_entry: VirtAddr,
}

The struct of Syscall which stores the information about making a syscall

Fields
thread: &'a CurrentThread

the thread making a syscall

thread_fn: ThreadFn

new thread function

syscall_entry: VirtAddr

the entry of current syscall

Implementations

impl Syscall<'_>

pub fn sys_getcwd(&self, buf: UserOutPtr<u8>, len: usize) -> SysResult

return a null-terminated string containing an absolute pathname that is the current working directory of the calling process.

buf – pointer to buffer to receive path

len – size of buf

pub fn sys_chdir(&self, path: UserInPtr<u8>) -> SysResult

Change the current directory.

path – pointer to string with name of path

pub fn sys_mkdir(&self, path: UserInPtr<u8>, mode: usize) -> SysResult

Make a directory.

path – pointer to string with directory name

mode – file system permissions mode

pub fn sys_mkdirat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
mode: usize
) -> SysResult

create directory relative to directory file descriptor

pub fn sys_rmdir(&self, path: UserInPtr<u8>) -> SysResult

Remove a directory.

path – pointer to string with directory name

pub fn sys_getdents64(
&self,
fd: FileDesc,
buf: UserOutPtr<u8>,
buf_size: usize
) -> SysResult

get directory entries TODO: get ino from dirent

fd – file describe

pub fn sys_link(
&self,
oldpath: UserInPtr<u8>,
newpath: UserInPtr<u8>
) -> SysResult

creates a new link (also known as a hard link) to an existing file.

pub fn sys_linkat(
&self,
olddirfd: FileDesc,
oldpath: UserInPtr<u8>,
newdirfd: FileDesc,
newpath: UserInPtr<u8>,
flags: usize
) -> SysResult

create file link relative to directory file descriptors If the pathname given in oldpath is relative, then it is interpreted relative to the directory referred to by the file descriptor olddirfd

pub fn sys_unlink(&self, path: UserInPtr<u8>) -> SysResult

delete name/possibly file it refers to If that name was the last link to a file and no processes have the file open, the file is deleted. If the name was the last link to a file but any processes still have the file open, the file will remain in existence until the last file descriptor referring to it is closed.

pub fn sys_unlinkat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
flags: usize
) -> SysResult

remove directory entry relative to directory file descriptor The unlinkat() system call operates in exactly the same way as either unlink or rmdir.

pub fn sys_rename(
&self,
oldpath: UserInPtr<u8>,
newpath: UserInPtr<u8>
) -> SysResult

change name/location of file

pub fn sys_renameat(
&self,
olddirfd: FileDesc,
oldpath: UserInPtr<u8>,
newdirfd: FileDesc,
newpath: UserInPtr<u8>
) -> SysResult

rename file relative to directory file descriptors

pub fn sys_readlink(
&self,
path: UserInPtr<u8>,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

read value of symbolic link

pub fn sys_readlinkat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

read value of symbolic link relative to directory file descriptor readlink() places the contents of the symbolic link path in the buffer base, which has size len TODO: recursive link resolution and loop detection

impl Syscall<'_>

pub fn sys_open(&self, path: UserInPtr<u8>, flags: usize, mode: usize) -> SysResult

Opens or creates a file, depending on the flags passed to the call. Returns an integer with the file descriptor.

pub fn sys_openat(
&self,
dir_fd: FileDesc,
path: UserInPtr<u8>,
flags: usize,
mode: usize
) -> SysResult

open file relative to directory file descriptor

pub fn sys_close(&self, fd: FileDesc) -> SysResult

Closes a file descriptor, so that it no longer refers to any file and may be reused.

pub fn sys_dup2(&self, fd1: FileDesc, fd2: FileDesc) -> SysResult

create a copy of the file descriptor oldfd.

pub fn sys_dup(&self, fd1: FileDesc) -> SysResult

create a copy of the file descriptor fd, and uses the lowest-numbered unused descriptor for the new descriptor.

pub fn sys_pipe(&self, fds: UserOutPtr<[i32; 2]>) -> SysResult

Creates a pipe, a unidirectional data channel that can be used for interprocess communication.

pub fn sys_pipe2(&self, fds: UserOutPtr<[i32; 2]>, flags: usize) -> SysResult

Creates a pipe, a unidirectional data channel that can be used for interprocess communication.

pub fn sys_flock(&mut self, fd: FileDesc, operation: usize) -> SysResult

apply or remove an advisory lock on an open file TODO: handle operation

impl Syscall<'_>

pub async fn sys_read(
&self,
fd: FileDesc,
base: UserOutPtr<u8>,
len: usize
) -> SysResult

Reads from a specified file using a file descriptor. Before using this call, you must first obtain a file descriptor using the opensyscall. Returns bytes read successfully.

fd – file descriptor

base – pointer to the buffer to fill with read contents

len – number of bytes to read

pub fn sys_write(&self, fd: FileDesc, base: UserInPtr<u8>, len: usize) -> SysResult

Writes to a specified file using a file descriptor. Before using this call, you must first obtain a file descriptor using the open syscall. Returns bytes written successfully.

fd – file descriptor

base – pointer to the buffer write

len – number of bytes to write

pub async fn sys_pread(
&self,
fd: FileDesc,
base: UserOutPtr<u8>,
len: usize,
offset: u64
) -> SysResult

read from or write to a file descriptor at a given offset reads up to count bytes from file descriptor fd at offset offset (from the start of the file) into the buffer starting at buf. The file offset is not changed.

pub fn sys_pwrite(
&self,
fd: FileDesc,
base: UserInPtr<u8>,
len: usize,
offset: u64
) -> SysResult

writes up to count bytes from the buffer starting at buf to the file descriptor fd at offset offset. The file offset is not changed.

pub async fn sys_readv(
&self,
fd: FileDesc,
iov_ptr: UserInPtr<IoVecOut>,
iov_count: usize
) -> SysResult

works just like read except that multiple buffers are filled. reads iov_count buffers from the file associated with the file descriptor fd into the buffers described by iov (“scatter input”)

pub fn sys_writev(
&self,
fd: FileDesc,
iov_ptr: UserInPtr<IoVecIn>,
iov_count: usize
) -> SysResult

works just like write except that multiple buffers are written out. writes iov_count buffers of data described by iov to the file associated with the file descriptor fd (“gather output”).

pub fn sys_lseek(&self, fd: FileDesc, offset: i64, whence: u8) -> SysResult

repositions the offset of the open file associated with the file descriptor fd to the argument offset according to the directive whence

pub fn sys_truncate(&self, path: UserInPtr<u8>, len: usize) -> SysResult

cause the regular file named by path to be truncated to a size of precisely length bytes.

pub fn sys_ftruncate(&self, fd: FileDesc, len: usize) -> SysResult

cause the regular file referenced by fd to be truncated to a size of precisely length bytes.

pub async fn sys_sendfile(
&self,
out_fd: FileDesc,
in_fd: FileDesc,
offset_ptr: UserInOutPtr<u64>,
count: usize
) -> SysResult

copies data between one file descriptor and another.

pub async fn sys_copy_file_range(
&self,
in_fd: FileDesc,
in_offset: UserInOutPtr<u64>,
out_fd: FileDesc,
out_offset: UserInOutPtr<u64>,
count: usize,
flags: usize
) -> SysResult

copies data between one file descriptor and anothe, read from specified offset and write new offset back

pub fn sys_sync(&self) -> SysResult

causes all buffered modifications to file metadata and data to be written to the underlying file systems.

pub fn sys_fsync(&self, fd: FileDesc) -> SysResult

transfers (“flushes”) all modified in-core data of (i.e., modified buffer cache pages for) the file referred to by the file descriptor fd to the disk device

pub fn sys_fdatasync(&self, fd: FileDesc) -> SysResult

is similar to fsync(), but does not flush modified metadata unless that metadata is needed

pub fn sys_ioctl(
&self,
fd: FileDesc,
request: usize,
arg1: usize,
arg2: usize,
arg3: usize
) -> SysResult

Set parameters of device files.

pub fn sys_fcntl(&self, fd: FileDesc, cmd: usize, arg: usize) -> SysResult

Manipulate a file descriptor.

cmd – cmd flag

arg – additional parameters based on cmd

pub fn sys_access(&self, path: UserInPtr<u8>, mode: usize) -> SysResult

Checks whether the calling process can access the file pathname

pub fn sys_faccessat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
mode: usize,
flags: usize
) -> SysResult

Check user’s permissions of a file relative to a directory file descriptor TODO: check permissions based on uid/gid

pub fn sys_utimensat(
&mut self,
dirfd: FileDesc,
pathname: UserInPtr<u8>,
times: UserInOutPtr<[TimeSpec; 2]>,
flags: usize
) -> SysResult

change file timestamps with nanosecond precision

pub fn sys_statfs(
&self,
path: UserInPtr<u8>,
buf: UserOutPtr<StatFs>
) -> SysResult

Get filesystem statistics (see linux man statfs(2)).

The statfs system call returns information about a mounted filesystem. path is the pathname of any file within the mounted filesystem. buf is a pointer to a StatFs structure.

pub fn sys_fstatfs(&self, fd: FileDesc, buf: UserOutPtr<StatFs>) -> SysResult

Get filesystem statistics (see linux man statfs(2)).

The fstatfs system call returns information about a mounted filesystem. fd is the descriptor referencing an open file. buf is a pointer to a StatFs structure.

impl Syscall<'_>

pub async fn sys_poll(
&mut self,
ufds: UserInOutPtr<PollFd>,
nfds: usize,
timeout_msecs: isize
) -> SysResult

Wait for some event on a file descriptor

pub async fn sys_ppoll(
&mut self,
ufds: UserInOutPtr<PollFd>,
nfds: usize,
timeout: UserInPtr<TimeSpec>
) -> SysResult

Wait for some event on a file descriptor

ppoll() allows an application to safely wait until either a file descriptor becomes ready or until a signal is caught

pub async fn sys_pselect6(
&mut self,
nfds: usize,
read: UserInOutPtr<u32>,
write: UserInOutPtr<u32>,
err: UserInOutPtr<u32>,
timeout: UserInPtr<TimeVal>,
_sigset: usize
) -> SysResult

similar to select, but have sigmask argument

pub async fn sys_select(
&mut self,
nfds: usize,
read: UserInOutPtr<u32>,
write: UserInOutPtr<u32>,
err: UserInOutPtr<u32>,
timeout: UserInPtr<TimeVal>
) -> SysResult

allow a program to monitor multiple file descriptors, waiting until one or more of the file descriptors become “ready” for some class of I/O operation.

A file descriptor is considered ready if it is possible to perform the corresponding I/O operation (e.g., read) without blocking.

impl Syscall<'_>

pub fn sys_lstat(
&self,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>
) -> SysResult

Works exactly like the stat syscall, but if the file in question is a symbolic link, information on the link is returned rather than its target.

path – full path to file

stat_ptr – pointer to stat buffer

pub fn sys_fstat(&self, fd: FileDesc, stat_ptr: UserOutPtr<Stat>) -> SysResult

Works exactly like the stat syscall except a file descriptor (fd) is provided instead of a path.

fd – file descriptor

stat_ptr – pointer to stat buffer

pub fn sys_fstatat(
&self,
dirfd: FileDesc,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>,
flags: usize
) -> SysResult

get file status relative to a directory file descriptor

pub fn sys_stat(
&self,
path: UserInPtr<u8>,
stat_ptr: UserOutPtr<Stat>
) -> SysResult

Returns information about a file in a structure named stat.

path – pointer to the name of the file

stat_ptr – pointer to the structure to receive file information

impl Syscall<'_>

Syscalls of inter-process communication and System V semaphore Set operation.

Menu

semget

semop

semctl

shmget

shmat

shmdt

shmctl

pub fn sys_semget(&self, key: usize, nsems: usize, flags: usize) -> SysResult

Get a System V semaphore set identifier (see linux man semget(2)).

The sys_semget system call returns the System V semaphore set identifier associated with the argument key. It may be used either to obtain the identifier of a previously created semaphore set (when flags is zero and key is not zero), or to create a new set.

A new set of nsems (number of semaphores) semaphores is created if key is zero or if no existing semaphore set is associated with key and IpcGetFlag::CREAT is specified in semflg.

If flags specifies both IpcGetFlag::CREAT and IpcGetFlag::EXCLUSIVE and a semaphore set already exists for key, then sys_semget fails with EEXIST. (This is analogous to the effect of the combination OpenFlags::CREATE | OpenFlags::EXCLUSIVE for sys_open.)

Upon creation, the least significant 9 bits of the argument flags define the permissions (for owner, group, and others) for the semaphore set. These bits have the same format, and the same meaning, as the mode argument of sys_open (though the execute permissions are not meaningful for semaphores, and write permissions mean permission to alter semaphore values).

When creating a new semaphore set, sys_semget initializes the set’s associated data structure, semid_ds (see sys_semctl), as follows:

sem_perm.cuid and sem_perm.uid are set to the effective user ID of the calling process.

sem_perm.cgid and sem_perm.gid are set to the effective group ID of the calling process.

The least significant 9 bits of sem_perm.mode are set to the least significant 9 bits of flags.

sem_nsems is set to the value of nsems.

sem_otime is set to 0.

sem_ctime is set to the current time.

The argument nsems can be 0 (a don’t care) when a semaphore set is not being created. Otherwise, nsems must be greater than 0 and less than or equal to the maximum number of semaphores per semaphore set (SEMMSL, constant 256).

If the semaphore set already exists, the permissions are verified.

pub async fn sys_semop(
&self,
id: usize,
ops: UserInPtr<SemBuf>,
num_ops: usize
) -> SysResult

System V semaphore operations (see linux man semop(2)).

semop performs operations on selected semaphores in the set indicated by id. An array [SemBuf; num_ops] pointed to by ops specifies an operation to be performed on a single semaphore. The declaration of SemBuf is like this:

struct SemBuf {
   num: u16,
   op: i16,
   flags: i16,
}

Flags recognized in SemBuf::flags are SemFlags::IPC_NOWAIT and SemFlags::SEM_UNDO. If an operation specifies SEM_UNDO, it will be automatically undone when the process terminates.

Each operation is performed on the SemBuf::num-th semaphore of the semaphore set, where the first semaphore of the set is numbered 0. There are two types of operation, distinguished by the value of SemBuf::op.

If op is +1, see acquire.

If op is -1, see release.

pub fn sys_semctl(&self, id: usize, num: usize, cmd: usize, arg: usize) -> SysResult

System V semaphore control operations (see linux man semctl(2)).

semctl performs the control operation specified by cmd on the System V semaphore set identified by id, or on the num-th semaphore of that set (The semaphores in a set are numbered starting at 0).

TODO

pub fn sys_shmget(&self, key: usize, size: usize, shmflg: usize) -> SysResult

Allocates a System V shared memory segment (see linux man shmget(2)).

shmget returns the identifier of the System V shared memory segment associated with the value of the argument key. Differ from linux, this syscall always create a new set.

pub fn sys_shmat(&self, id: usize, addr: VirtAddr, shmflg: usize) -> SysResult

System V shared memory operations (see linux man shmat(2)).

shmat attaches the System V shared memory segment identified by id to the address space of the calling process. The attaching address is specified by addr. If addr is zero, the system chooses a suitable page-aligned address to attach the segment.

pub fn sys_shmdt(&self, id: usize, addr: VirtAddr, shmflg: usize) -> SysResult

System V shared memory operations (see linux man shmdt(2)).

shmdt detaches the shared memory segment located at the address specified by addr from the address space of the calling process. The to-be-detached segment must be currently attached with addr equal to the value returned by the attaching sys_shmat call.

pub fn sys_shmctl(&self, id: usize, cmd: usize, buffer: usize) -> SysResult

System V shared memory operations (see linux man shmctl(2)).

performs the control operation specified by cmd on the shared memory segment whose identifier is given in id

impl Syscall<'_>

pub fn sys_arch_prctl(&mut self, code: i32, addr: usize) -> SysResult

set architecture-specific thread state for x86_64 currently

pub fn sys_uname(&self, buf: UserOutPtr<u8>) -> SysResult

get name and information about current kernel

pub fn sys_sysinfo(&mut self, sys_info: UserOutPtr<SysInfo>) -> SysResult

provides a simple way of getting overall system statistics

pub async fn sys_futex(
&self,
uaddr: usize,
op: u32,
val: u32,
val2: usize,
uaddr2: usize,
_val3: u32
) -> SysResult

provides a method for waiting until a certain condition becomes true.

uaddr - points to the futex word.

op - the operation to perform on the futex

val - a value whose meaning and purpose depends on op

val2 - provides a timeout for the attempt or acts as val2 when op is REQUEUE

uaddr2 - when op is REQUEUE, points to the target futex

_val3 - is not used

pub fn sys_prlimit64(
&mut self,
pid: usize,
resource: usize,
new_limit: UserInPtr<RLimit>,
old_limit: UserOutPtr<RLimit>
) -> SysResult

Combines and extends the functionality of setrlimit() and getrlimit()

pub fn sys_getrandom(
&mut self,
buf: UserOutPtr<u8>,
len: usize,
flag: u32
) -> SysResult

fills the buffer pointed to by buf with up to buflen random bytes.

buf - buffer that needed to fill

buflen - length of buffer

flag - a bit mask that can contain zero or more of the following values ORed together:

GRND_RANDOM

GRND_NONBLOCK

returns the number of bytes that were copied to the buffer buf.

impl Syscall<'_>

pub fn sys_socket(
&mut self,
domain: usize,
_type: usize,
protocol: usize
) -> SysResult

creates an endpoint for communication and returns a file descriptor that refers to that endpoint.

pub async fn sys_connect(
&mut self,
sockfd: usize,
addr: UserInPtr<SockAddr>,
addrlen: usize
) -> SysResult

connects the socket referred to by the file descriptor sockfd to the address specified by addr.

pub fn sys_setsockopt(
&mut self,
sockfd: usize,
level: usize,
optname: usize,
optval: UserInPtr<u8>,
optlen: usize
) -> SysResult

set options for the socket referred to by the file descriptor sockfd.

pub fn sys_getsockopt(
&mut self,
sockfd: usize,
level: usize,
optname: usize,
optval: UserOutPtr<u32>,
optlen: UserOutPtr<u32>
) -> SysResult

get options for the socket referred to by the file descriptor sockfd.

pub fn sys_sendto(
&mut self,
sockfd: usize,
buf: UserInPtr<u8>,
len: usize,
flags: usize,
dest_addr: UserInPtr<SockAddr>,
addrlen: usize
) -> SysResult

transmit a message to another socket

pub async fn sys_recvfrom(
&mut self,
sockfd: usize,
buf: UserOutPtr<u8>,
len: usize,
flags: usize,
src_addr: UserOutPtr<SockAddr>,
addrlen: UserInOutPtr<u32>
) -> SysResult

receive messages from a socket

pub async fn sys_recvmsg(
&mut self,
sockfd: usize,
msg: UserInOutPtr<MsgHdr>,
flags: usize
) -> SysResult

receive messages from a socket

pub fn sys_bind(
&mut self,
sockfd: usize,
addr: UserInPtr<SockAddr>,
addrlen: usize
) -> SysResult

assigns the address specified by addr to the socket referred to by the file descriptor sockfd

pub fn sys_listen(&mut self, sockfd: usize, backlog: usize) -> SysResult

marks the socket referred to by sockfd as a passive socket, that is, as a socket that will be used to accept incoming connection

pub fn sys_shutdown(&mut self, sockfd: usize, howto: usize) -> SysResult

shutdown a socket

pub async fn sys_accept(
&mut self,
sockfd: usize,
addr: UserOutPtr<SockAddr>,
addrlen: UserInOutPtr<u32>
) -> SysResult

accept() is used with connection-based socket types (SOCK_STREAM, SOCK_SEQPACKET). It extracts the first connection request on the queue of pending connections for the listening socket, sockfd, creates a new connected socket, and returns a new file descriptor referring to that socket. The newly created socket is not in the listening state. The original socket sockfd is unaffected by this call.

pub fn sys_getsockname(
&mut self,
sockfd: usize,
addr: UserOutPtr<SockAddr>,
addrlen: UserInOutPtr<u32>
) -> SysResult

returns the current address to which the socket sockfd is bound, in the buffer pointed to by addr.

pub fn sys_getpeername(
&mut self,
sockfd: usize,
addr: UserOutPtr<SockAddr>,
addrlen: UserInOutPtr<u32>
) -> SysResult

returns the address of the peer connected to the socket sockfd, in the buffer pointed to by addr.

impl Syscall<'_>

pub fn sys_rt_sigaction(
&self,
signum: usize,
act: UserInPtr<SignalAction>,
oldact: UserOutPtr<SignalAction>,
sigsetsize: usize
) -> SysResult

Used to change the action taken by a process on receipt of a specific signal.

pub fn sys_rt_sigprocmask(
&mut self,
how: i32,
set: UserInPtr<Sigset>,
oldset: UserOutPtr<Sigset>,
sigsetsize: usize
) -> SysResult

Used to fetch and/or change the signal mask of the calling thread

pub fn sys_sigaltstack(
&self,
ss: UserInPtr<SignalStack>,
old_ss: UserOutPtr<SignalStack>
) -> SysResult

Allows a process to define a new alternate signal stack and/or retrieve the state of an existing alternate signal stack

pub fn sys_kill(&self, pid: isize, signum: usize) -> SysResult

Send a signal to a process specified by pid TODO: support all the arguments

pub fn sys_tkill(&mut self, tid: usize, signum: usize) -> SysResult

Send a signal to a thread specified by tid

pub fn sys_tgkill(&mut self, tgid: usize, tid: usize, signum: usize) -> SysResult

Send a signal to a thread specified by tgid (i.e., process) and pid Note: the job of the target process should be the same as the calling thread

pub fn sys_rt_sigreturn(&mut self) -> SysResult

Return from handling some signal

impl Syscall<'_>

Syscalls for process.

Menu

fork

vfork

clone

wait4

execve

gettid

getpid

getppid

exit

exit_group

nanosleep

set_tid_address

pub fn sys_fork(&self) -> SysResult

fork creates a new process by duplicating the calling process (see linux man fork(2)). The new process is referred to as the child process. The calling process is referred to as the parent process.

The child process and the parent process run in separate memory spaces. At the time of fork both memory spaces have the same content. Memory writes, file mappings (Self::sys_mmap) and unmappings (Self::sys_munmap) performed by one of the processes do not affect the other.

The child process is an exact duplicate of the parent process except for the following points:

The child has its own unique process ID, and this PID does not match the ID of any existing process.

The child’s parent process ID is the same as the parent’s process ID.

Process resource utilizations (Self::sys_getrusage) and CPU time counters (Self::sys_times) are reset to zero in the child.

The child does not inherit semaphore adjustments from its parent (Self::sys_semop).

The child does not inherit process-associated record locks from its parent (Self::sys_fcntl). (On the other hand, it does inherit Self::sys_fcntl open file description locks and Self::sys_flock locks from its parent.)

Note the following further points:

The child process is created with a single thread—the one that called fork(). The entire virtual address space of the parent is replicated in the child, including the states of mutexes and condition variables.

After a fork in a multithreaded program, the child can safely call only async-signal-safe functions until such time as it calls Self::sys_execve.

The child inherits copies of the parent’s set of open file descriptors. Each file descriptor in the child refers to the same open file description (see Self::sys_open) as the corresponding file descriptor in the parent. This means that the two file descriptors share open file status flags and file offset.

pub async fn sys_vfork(&self) -> SysResult

sys_vfork, just like Self::sys_fork, creates a child process of the calling process (see linux man vfork(2)). For details, see Self::sys_fork.

sys_vfork differs from Self::sys_fork in that the calling thread is suspended until the child terminates (either normally, by calling Self::sys_exit, or abnormally, after delivery of a fatal signal), or it makes a call to Self::sys_execve.

pub fn sys_clone(
&self,
flags: usize,
newsp: usize,
parent_tid: UserOutPtr<i32>,
newtls: usize,
child_tid: UserOutPtr<i32>
) -> SysResult

sys_clone create a new thread in the current process. The new thread’s stack pointer will be set to newsp, and thread pointer will be set to newtls. The child TID will be stored at both parent_tid and child_tid.

NOTE! This system call is not exactly the same as clone in Linux.

NOTE! This is partially implemented for musl only.

pub async fn sys_wait4(
&self,
pid: i32,
wstatus: UserOutPtr<i32>,
options: u32
) -> SysResult

sys_wait4 suspends execution of the calling thread until a child specified by pid argument has changed state (see linux man wait4(2)). By default, sys_wait4 waits only for terminated children, but this behavior is modifiable via the options argument, as described below.

The value of pid can be:

-1: meaning wait for any child process.

0: meaning wait for any child process whose process group ID is equal to that of the calling process at the time of the call to sys_wait4.

>0: meaning wait for the child whose process ID is equal to the value of pid.

The value of options is an OR of zero or more of the following constants:

NOHANG = 0x000_0001;

TODO

STOPPED = 0x000_0002;

TODO

EXITED = 0x000_0004;

TODO

CONTINUED = 0x000_0008;

TODO

NOWAIT = 0x100_0000;

TODO

On success, returns the process ID of the child whose state has changed; if NOHANG flag was specified and one or more child(ren) specified by pid exist, but have not yet changed state, then 0 is returned. On failure, -1 is returned.

pub fn sys_execve(
&mut self,
path: UserInPtr<u8>,
argv: UserInPtr<UserInPtr<u8>>,
envp: UserInPtr<UserInPtr<u8>>
) -> SysResult

sys_execve executes the program referred to by path (see linux man execve(2)). This causes the program that is currently being run by the calling process to be replaced with a new program, with newly initialized stack, heap, and (initialized and uninitialized) data segments.

path argument must be a binary executable file.

argv is an array of argument strings passed to the new program. By convention, the first of these strings (i.e., argv[0]) should contain the filename associated with the file being executed.

envp is an array of strings, conventionally of the form key=value, which are passed as environment to the new program.

NOTE! Differ from linux, argv & envp can not be NULL.

NOTE! For multi-thread programs, A call to any exec function from a process with more than one thread shall result in all threads being terminated and the new executable image being loaded and executed.

pub fn sys_gettid(&self) -> SysResult

sys_gettid returns the caller’s thread ID (TID) (see linux man gettid(2)). In a single-threaded process, the thread ID is equal to the process ID (PID, as returned by Self::sys_getpid). In a multithreaded process, all threads have the same PID, but each one has a unique TID.

pub fn sys_getpid(&self) -> SysResult

sys_getpid returns the process ID (PID) of the calling process (see linux man getpid(2)).

pub fn sys_getppid(&self) -> SysResult

sys_getppid returns the process ID of the parent of the calling process (see linux man getppid(2)). This will be either the ID of the process that created this process using fork(), or, if that process has already terminated, 0.

pub fn sys_exit(&mut self, exit_code: i32) -> SysResult

sys_exit system call terminates only the calling thread (see linux man _exit(2), this syscall is same as a raw _exit in glibc), and actions such as reparenting child processes or sending SIGCHLD to the parent process are performed only if this is the last thread in the thread group.

pub fn sys_exit_group(&mut self, exit_code: i32) -> SysResult

sys_exit_group is equivalent to Self::sys_exit except that it terminates not only the calling thread (see linux man exit_group(2), but all threads in the calling process’s thread group. As a result, the entire calling process will exit.

pub async fn sys_nanosleep(&self, req: UserInPtr<TimeSpec>) -> SysResult

Allows the calling thread to sleep for an interval specified with nanosecond precision (see linux man nanosleep(2).

nanosleep suspends the execution of the calling thread until either at least the time specified in req has elapsed, or the delivery of a signal that triggers the invocation of a handler in the calling thread or that terminates the process.

To represent a duration, see TimeSpec.

pub fn sys_set_tid_address(&self, tidptr: UserOutPtr<i32>) -> SysResult

set_tid_address sets the clear_child_tid value for the calling thread to tidptr, and return the caller’s thread ID (see linux man set_tid_address(2).

pub fn sys_get_robust_list(
&self,
pid: i32,
head_ptr: UserOutPtr<UserOutPtr<RobustList>>,
len_ptr: UserOutPtr<usize>
) -> SysResult

Get robust list.

pub fn sys_set_robust_list(
&self,
head: UserInPtr<RobustList>,
len: usize
) -> SysResult

Set robust list.

impl Syscall<'_>

pub fn sys_clock_gettime(
&self,
clock: usize,
buf: UserOutPtr<TimeSpec>
) -> SysResult

finds the resolution (precision) of the specified clock clockid, and, if buffer is non-NULL, stores it in the struct timespec pointed to by buffer

pub fn sys_gettimeofday(
&mut self,
tv: UserOutPtr<TimeVal>,
tz: UserInPtr<u8>
) -> SysResult

get the time with second and microseconds

pub fn sys_time(&mut self, time: UserOutPtr<u64>) -> SysResult

get time in seconds

pub fn sys_block_in_kernel(&self) -> SysResult

JUST FOR TEST, DO NOT USE IT

pub fn sys_getrusage(
&mut self,
who: usize,
rusage: UserOutPtr<RUsage>
) -> SysResult

get resource usage currently only support ru_utime and ru_stime:

ru_utime: user CPU time used

ru_stime: system CPU time used

pub fn sys_times(&mut self, buf: UserOutPtr<Tms>) -> SysResult

stores the current process times in the struct tms that buf points to

pub async fn sys_clock_nanosleep(
&self,
clockid: usize,
flags: usize,
req: UserInPtr<TimeSpec>,
rem: UserOutPtr<TimeSpec>
) -> SysResult

clock nanosleep

impl Syscall<'_>

Syscalls for virtual memory.

Menu

mmap

mprotect

munmap

pub async fn sys_mmap(
&self,
addr: usize,
len: usize,
prot: usize,
flags: usize,
fd: FileDesc,
offset: u64
) -> SysResult

Map files or devices into memory (see linux man mmap(2)).

sys_mmap creates a new mapping in the virtual address space of the calling process.

The starting address for the new mapping is specified in addr.

The len argument specifies the length of the mapping (which must be greater than 0).

Arguments fd and offset specifies mapping file descriptor and offset in the file.

The prot argument describes the desired memory protection of the mapping (and must not conflict with the open mode of the file). It is either 0 or the bitwise OR of one or more of the following flags:

MmapProt::READ

Pages may be read

MmapProt::WRITE

Pages may be written

MmapProt::EXEC

Pages may be executed

The flags argument determines whether updates to the mapping are visible to other processes mapping the same region, and whether updates are carried through to the underlying file. This behavior is determined by including exactly one of the following values:

MmapFlags::SHARED

Share this mapping. Updates to the mapping are visible to other processes mapping the same region, and (in the case of file-backed mappings) are carried through to the underlying file. (To precisely control when updates are carried through to the underlying file requires the use of msync, which has not been implemented in zcore).

MmapFlags::PRIVATE

Create a private copy-on-write mapping. Updates to the mapping are not visible to other processes mapping the same file, and are not carried through to the underlying file. It is unspecified whether changes made to the file after the sys_mmap call are visible in the mapped region.

MmapFlags::FIXED

Don’t interpret addr as a hint: place the mapping at exactly that address. addr must be suitably aligned: for most architectures a multiple of the page size is sufficient; however, some architectures may impose additional restrictions. If the memory region specified by addr and len overlaps pages of any existing mapping(s), then the overlapped part of the existing mapping(s) will be discarded. If the specified address cannot be used, sys_mmap will fail.

MmapFlags::ANONYMOUS

The mapping is not backed by any file; its contents are initialized to zero. Both fd and offset arguments are ignored. The use of MmapFlags::ANONYMOUS in conjunction with MmapFlags::SHARED causes an EINVAL to be returned.

pub fn sys_mprotect(&self, addr: usize, len: usize, prot: usize) -> SysResult

Set protection on a region of memory (see linux man mprotect(2)).

NOTE! This syscall is now unimplemented. Calling it always return Ok(0).

sys_mprotect changes the access protections for the calling process’s memory pages containing any part of the address range in the interval [addr, addr+len-1]. addr must be aligned to a page boundary.

If the calling process tries to access memory in a manner that violates the protections, then the kernel generates a SIGSEGV signal for the process.

prot is a combination of the following access flags: 0 or a bitwise-or of the other values in the following list:

MmapProt::READ

The memory can be read.

MmapProt::WRITE

The memory can be modified.

MmapProt::EXEC

The memory can be executed.

If prot is 0, the memory cannot be accessed at all.

pub fn sys_munmap(&self, addr: usize, len: usize) -> SysResult

Unmap files or devices into memory (see linux man munmap(2)).

Deletes the mappings for the specified address range, and causes further references to addresses within the range to generate invalid memory references.

The sys_munmap system call deletes the mappings for the specified address range, and causes further references to addresses within the range to generate invalid memory references. The region is also automatically unmapped when the process is terminated. On the other hand, closing the file descriptor does not unmap the region.

Both addr and len must be aligned to the page size, additionally, len must greater than 0. Otherwise, an EINVAL is returned.

impl Syscall<'_>

pub async fn syscall(&mut self, num: u32, args: [usize; 6]) -> isize

syscall entry function